# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",
                "description": "Get users details",
                "operationId": "GetUsersDetails",
                "responses": {
                    "200": {
                        "description": "Fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register a user",
                "description": "Register a user",
                "operationId": "registerUsers",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "role",
                                    "province",
                                    "identificationnumber"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "role": {
                                        "type": "text"
                                    },
                                    "province": {
                                        "type": "text"
                                    },
                                    "district": {
                                        "type": "text"
                                    },
                                    "sector": {
                                        "type": "text"
                                    },
                                    "identificationnumber": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/districts": {
            "get": {
                "tags": [
                    "Districts"
                ],
                "summary": "Get All Districts",
                "description": "Get Districts details",
                "operationId": "GetDistrictsDetails",
                "responses": {
                    "200": {
                        "description": "Fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Districts"
                ],
                "summary": "Register a district",
                "description": "Register a district",
                "operationId": "registerDistricts",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "provinceId": {
                                        "type": "text"
                                    },
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/districts/{district}": {
            "get": {
                "tags": [
                    "Districts"
                ],
                "summary": "Get A Single District",
                "description": "Get Districts details",
                "operationId": "GetdistrictDetails",
                "parameters": [
                    {
                        "name": "district",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Districts"
                ],
                "summary": "Update a district",
                "description": "update a district",
                "operationId": "UpdateDistrict",
                "parameters": [
                    {
                        "name": "district",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "privinceId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "provinceId": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Districts"
                ],
                "summary": "Delete A Single District",
                "description": "Delete District details",
                "operationId": "DeleteDistrictDetails",
                "parameters": [
                    {
                        "name": "district",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}